+++ T2TScripts/functions/Import-T2TAttributes.ps1
@@ -56,7 +56,7 @@
 
     .NOTES
         Title: Import-T2TAttributes.ps1
-        Version: 1.2.1
+        Version: 2.0.0
         Date: 2021.01.03
         Author: Denis Vilaca Signorelli (denis.signorelli@microsoft.com)
         Contributors: Agustin Gallegos (agustin.gallegos@microsoft.com)
@@ -321,7 +321,7 @@
 
     Write-PSFMessage -Level Output -Message "The import is completed. Please confirm that all users are correctly created before enable the Azure AD Sync Cycle."
     Write-PSFMessage -Level Output -Message "You can re-enable Azure AD Connect using the following cmdlet: 'Set-ADSyncScheduler -SyncCycleEnabled 1'"
-    Remove-Variable * -ErrorAction SilentlyContinue
     Get-PSSession | Remove-PSSession
+    Remove-Variable * -ErrorAction SilentlyContinue
 
 }